Extensive overhaul of KStars header files:
+ Forward-declare classes wherever possible (i.e.,
use "class Foo;" instead of "#include <foo.h>")

+ Improve DOxygen comments

+ Remove some unused functions

+ Fix some compiler warnings

The forward-declarations caused a lot of changes.
Many cpp files were depending on
remotely-included headers (i.e, it needs a.h, but
it uses '#include "b.h"'; b.h has '#include "c.h";
c.h has '#include "a.h"'...yuck).  We should now
be much closer to the ideal where each file
includes the headers of the classes that it needs.

Almost all of the header files now have good DOxygen
API comments.  The few classes which don't are ones
that I don't feel qualified to document (i.e., INDI).

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=302474

+++ kstars/main.cpp
@@ -22,6 +22,10 @@
 #include <kdebug.h>
 
 #include "kstars.h"
+#include "kstarsdata.h"
+#include "skymap.h"
+#include "simclock.h"
+#include "ksnumbers.h"
 #include "Options.h"
 
 #define KSTARS_VERSION "1.0"
@@ -99,7 +103,7 @@ int main(int argc, char *argv[])
 		dat->colorScheme()->loadFromConfig( kapp->config() );
 
 		//reset clock now that we have a location:
-		dat->clock()->setUTC( QDateTime::currentDateTime().addSecs( -3600 * dat->geo()->TZ() ) );
+		dat->clock()->setUTC( QDateTime::currentDateTime().addSecs( int( -3600 * dat->geo()->TZ() ) ) );
 
 		KSNumbers num( dat->clock()->JD() );
 		dat->initGuides(&num);
